creando stripe 2

+++ src/main/java/com/jgm/paladohorweb/tour/dto/request/TourRequestDTO.java
@@ -22,5 +22,5 @@ public class TourRequestDTO {
     private String imagenUrl;
     private Double latitud;
     private Double longitud;
-    private BigDecimal precio;
+    private Long precio;
 }
